Starter code.
I unpacked the handout directory and then moved the .git folder and README out
of the 440_P2 folder into P2/go/src/P2-f12/ directory inside the code tree.

Now only changes to the contrib folder will be visible to either of us and the
bin or pkg directories are separate.