Start libstore implementation
I wrote the NewLibstore function. It just contacts the master server and
fetches the list of storage servers (trying multiple times if they aren't
ready).

At this point, we should ignore all of the servers except the master when doing
storage, then move on to routing requests once the basic code works.