Modify libstore cache implementation.
I modified the struct to keep a list of query times since keeping the first
query time doesn't let us know recent behavior once its older than
QUERY_CACHE_SECONDS.

The libstore cache entry keeps a linked list of time.Time objects, with the
oldest at the head. When a query is made, we clear the list of older times and
if the number remaining is at least QUERY_CACHE_THRESH, then we can go ahead
and ask for the lease.

I also uncommented the check for the force lease flag since it doesn't do
anything if we don't pass the flag on the command-line.